Add spring integration tests
Test that the Spring context is correctly setup and the controller calls our delegate. This is pretty much just a smoke test to check that the http layer is wired correctly.

This way, the next tests I'm going to write will be either simple self-contained unit tests, or DB integration tests. Not touching Spring anymore.

I took the opportunity to clean up the API and the code generator a bit.